Add Entity class
